pyproject.toml
[tool.poetry]
name = "tg-notifications-admin"
version = "0.1.0"
description = ""
authors = ["Maksim Panin <me@naxrevlis.com>"]
readme = "README.md"
packages = [{include = "tg_notifications_admin"}]

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.2"
uvicorn = "^0.18.3"
tortoise = "^0.1.1"
fastapi = "^0.85.0"
pydantic = "^1.10.2"
aerich = "^0.6.3"
asyncpg = "^0.26.0"
python-dotenv = "^0.21.0"
passlib = "^1.7.4"
bcrypt = "^4.0.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "tests", "tg_notifications_admin"]
